push & update to version 0.11.3? I don't know.
Idk if it's 0.11.3.

Changes made:
Title has been changed. "text editor" was replaced with "writing program".

`textarea` was changed to a `div` with `contenteditable`, along with the appropriate CSS and JS changes also.

Add working placeholder text with the `data-text` tag, along with the appropriate code. which meant for the removal of the existing placeholder.

neuomorphism effects were changed/tweaked slightly.

and the current version is printed to the console log.